Move formatDate.js
